chore: rename critical-log-review to error-log-review across code, docs, and jobs
