--- tools/critical-log-review/go.mod
@@ -1,21 +1,0 @@
-module github.com/PingCAP-QE/ci/tools/critical-log-review
-
-go 1.20
-
-require (
-	github.com/google/go-github/v53 v53.2.0
-	golang.org/x/oauth2 v0.11.0
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/ProtonMail/go-crypto v0.0.0-20230217124315-7d5c6f04bbb8 // indirect
-	github.com/cloudflare/circl v1.3.3 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
-	golang.org/x/crypto v0.12.0 // indirect
-	golang.org/x/net v0.14.0 // indirect
-	golang.org/x/sys v0.11.0 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
